update tests for copy

+++ bento_backup_utils/tests/fs.rs
@@ -1,7 +1,7 @@
 // cargo build && cargo test -- --test-threads=1 --nocapture
 
 use std::error::Error;
-use std::path::{Path, PathBuf};
+use std::path::{Path};
 use std::panic;
 use std::fs as rust_fs;
 
@@ -13,7 +13,9 @@ const TEST_FOLDER: &'static str = "./tmp";
 
 #[allow(dead_code)]
 fn setup() {
-    rust_fs::remove_dir_all(TEST_FOLDER).unwrap();
+    if Path::new(TEST_FOLDER).exists() {
+        rust_fs::remove_dir_all(TEST_FOLDER).unwrap();
+    }
 }
 
 #[allow(dead_code)]
@@ -49,28 +51,21 @@ where
 #[test]
 fn test_file_eq() {
     run_test(||{
-        let mut test_file1 = PathBuf::from(TEST_FOLDER);
-        test_file1.push("test_file_eq");
-        test_file1.push("test1.txt");
-        let mut test_file2 = PathBuf::from(TEST_FOLDER);
-        test_file2.push("test_file_eq");
-        test_file2.push("test2.txt");
-        fs_extra::dir::create_all(test_file1.parent().unwrap(), true).unwrap();
-        fs_extra::dir::create_all(test_file2.parent().unwrap(), true).unwrap();
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_file_eq"), false).unwrap();
 
         let content1 = "content";
-        fs_extra::file::write_all(&test_file1, &content1).unwrap();
-        assert!(test_file1.exists());
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_file_eq/test1.txt"), &content1).unwrap();
+        assert!(Path::new(TEST_FOLDER).join("test_file_eq/test1.txt").exists());
         let content2 = "content";
-        fs_extra::file::write_all(&test_file2, &content2).unwrap();
-        assert!(test_file2.exists());
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_file_eq/test2.txt"), &content2).unwrap();
+        assert!(Path::new(TEST_FOLDER).join("test_file_eq/test2.txt").exists());
 
-        let read1 = fs_extra::file::read_to_string(&test_file1).unwrap();
+        let read1 = fs_extra::file::read_to_string(&Path::new(TEST_FOLDER).join("test_file_eq/test1.txt")).unwrap();
         assert_eq!(read1, content1);
-        let read2 = fs_extra::file::read_to_string(&test_file2).unwrap();
+        let read2 = fs_extra::file::read_to_string(&Path::new(TEST_FOLDER).join("test_file_eq/test2.txt")).unwrap();
         assert_eq!(read2, content2);
 
-        assert!(file_eq(test_file1, test_file2).unwrap());
+        assert!(file_eq(Path::new(TEST_FOLDER).join("test_file_eq/test1.txt"), Path::new(TEST_FOLDER).join("test_file_eq/test2.txt")).unwrap());
     })
 }
 
@@ -78,31 +73,16 @@ fn test_file_eq() {
 #[test]
 fn test_copy_no_files() {
     run_test(||{
-        let file_list = Vec::new();
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_no_files");
-
-        // Get base dir
-        let mut base_dir = PathBuf::from(TEST_FOLDER);
-        base_dir.push("test_copy_no_files");
-        base_dir.push("src");
-        assert!(!target_dir.exists()); // check there's a folder
-
         // Check if target_dir folder doesn't exist
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_no_files");
-        target_dir.push("target");
-        assert!(!target_dir.exists()); // check there's a folder
+        assert!(!Path::new(TEST_FOLDER).join("test_copy_no_files/target").exists());
 
         // run copy and check ok
-        let result = fs::copy(file_list, &base_dir, &target_dir);
+        let file_list = Vec::new();
+        let result = fs::copy(file_list, &Path::new(TEST_FOLDER).join("test_copy_no_files/src"), &Path::new(TEST_FOLDER).join("test_copy_no_files/target"));
         assert!(result.is_ok());
 
         // Check if target_dir folder exists
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_no_files");
-        target_dir.push("target");
-        assert!(target_dir.exists()); // check there's a folder
+        assert!(Path::new(TEST_FOLDER).join("test_copy_no_files/target").exists()); // check there's a folder
     })
 }
 
@@ -111,63 +91,33 @@ fn test_copy_no_files() {
 #[test]
 fn test_copy_to_existing_folder() {
     run_test(||{
-        // Get base dir
-        let mut base_dir = PathBuf::from(TEST_FOLDER);
-        base_dir.push("test_copy_to_existing_folder");
-        base_dir.push("src");
-
-        let file_list = Vec::new();
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_to_existing_folder");
-        target_dir.push("target");
-
         // create folder and check if there's no folder
-        assert!(!target_dir.exists());
-        fs_extra::dir::create_all(target_dir, true).unwrap();
+        assert!(!Path::new(TEST_FOLDER).join("test_copy_to_existing_folder/target").exists());
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_to_existing_folder/target"), false).unwrap();
 
         // check there's a folder
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_to_existing_folder");
-        target_dir.push("target");
-        assert!(target_dir.exists());
+        assert!(Path::new(TEST_FOLDER).join("test_copy_to_existing_folder/target").exists());
 
         // run copy and expect an error
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_copy_to_existing_folder");
-        let result = fs::copy(file_list, &base_dir, &target_dir);
+        let file_list = Vec::new();
+        let result = fs::copy(file_list, &Path::new(TEST_FOLDER).join("test_copy_to_existing_folder/src"), &Path::new(TEST_FOLDER).join("test_copy_to_existing_folder/target"));
         assert!(result.is_err());
     })
 }
 
 #[test]
-fn test_sample_copy1() {
+fn test_simple_copy() {
     run_test(||{
-        // Get base dir
-        let mut base_dir = PathBuf::from(TEST_FOLDER);
-        base_dir.push("test_sample_copy1");
-        base_dir.push("src");
-
         // create source directory
-        let mut src = PathBuf::from(TEST_FOLDER);
-        src.push("test_sample_copy1");
-        src.push("src");
-        fs_extra::dir::create_all(src, true).unwrap();
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_simple_copy/src"), false).unwrap();
 
         // create test1.txt
         let content1 = "content";
-        let mut test1 = PathBuf::from(TEST_FOLDER);
-        test1.push("test_sample_copy1");
-        test1.push("src");
-        test1.push("test1.txt");
-        fs_extra::file::write_all(&test1, &content1).unwrap();
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_simple_copy/src/test1.txt"), &content1).unwrap();
 
         // create test2.txt
         let content2 = "content";
-        let mut test2 = PathBuf::from(TEST_FOLDER);
-        test2.push("test_sample_copy1");
-        test2.push("src");
-        test2.push("test2.txt");
-        fs_extra::file::write_all(&test2, &content2).unwrap();
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_simple_copy/src/test2.txt"), &content2).unwrap();
 
         // create file list
         let mut file_list = Vec::new();
@@ -175,24 +125,108 @@ fn test_sample_copy1() {
         let path2 = "test2.txt";
         file_list.push(path2.to_string());
         file_list.push(path1.to_string());
-        println!("file_list = {:?}", file_list);
 
         // run copy and expect an error
-        let mut target_dir = PathBuf::from(TEST_FOLDER);
-        target_dir.push("test_sample_copy1");
-        target_dir.push("target");
-        let result = fs::copy(file_list, &base_dir, &target_dir);
+        let result = fs::copy(file_list, &Path::new(TEST_FOLDER).join("test_simple_copy/src"), &Path::new(TEST_FOLDER).join("test_simple_copy/target"));
         assert!(result.is_ok());
 
         // check target files exist
-        let path1_target = Path::new(TEST_FOLDER).join("test_sample_copy1/target/test1.txt");
-        let path2_target = Path::new(TEST_FOLDER).join("test_sample_copy1/target/test2.txt");
+        let path1_target = Path::new(TEST_FOLDER).join("test_simple_copy/target/test1.txt");
+        let path2_target = Path::new(TEST_FOLDER).join("test_simple_copy/target/test2.txt");
         assert!(path1_target.exists());
         assert!(path2_target.exists());
 
         // compare contents
-        let path1_src = Path::new(TEST_FOLDER).join("test_sample_copy1/src/test1.txt");
-        let path2_src = Path::new(TEST_FOLDER).join("test_sample_copy1/src/test2.txt");
+        let path1_src = Path::new(TEST_FOLDER).join("test_simple_copy/src/test1.txt");
+        let path2_src = Path::new(TEST_FOLDER).join("test_simple_copy/src/test2.txt");
+        assert!(file_eq(path1_src, path1_target).unwrap());
+        assert!(file_eq(path2_src, path2_target).unwrap());
+    })
+}
+
+#[test]
+fn test_copy_nested_folder1() {
+    run_test(||{
+        // create source directory
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src"), false).unwrap();
+
+        // create test1.txt
+        let content1 = "content";
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src/folder1"), false).unwrap();
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src/folder1/test1.txt"), &content1).unwrap();
+
+        // create test2.txt
+        let content2 = "content";
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src/test2.txt"), &content2).unwrap();
+
+        // create file list
+        let mut file_list = Vec::new();
+        let path1 = "folder1/test1.txt";
+        let path2 = "test2.txt";
+        file_list.push(path2.to_string());
+        file_list.push(path1.to_string());
+
+        // run copy and expect an error
+        let result = fs::copy(file_list, &Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src"), &Path::new(TEST_FOLDER).join("test_copy_nested_folder1/target"));
+        assert!(result.is_ok());
+
+        // check target files exist
+        let path1_target = Path::new(TEST_FOLDER).join("test_copy_nested_folder1/target/folder1/test1.txt");
+        let path2_target = Path::new(TEST_FOLDER).join("test_copy_nested_folder1/target/test2.txt");
+        assert!(path1_target.exists());
+        assert!(path2_target.exists());
+
+        // compare contents
+        let path1_src = Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src/folder1/test1.txt");
+        let path2_src = Path::new(TEST_FOLDER).join("test_copy_nested_folder1/src/test2.txt");
+        assert!(file_eq(path1_src, path1_target).unwrap());
+        assert!(file_eq(path2_src, path2_target).unwrap());
+    })
+}
+
+#[test]
+fn test_copy_nested_folder2() {
+    run_test(||{
+        // create source directory
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src"), false).unwrap();
+
+        // create test1.txt
+        let content1 = "content";
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder1/folder1_1"), false).unwrap();
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder1/folder1_1/test1.txt"), &content1).unwrap();
+
+        // create test2.txt
+        let content2 = "content";
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder2"), false).unwrap();
+        fs_extra::file::write_all(&Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder2/test2.txt"), &content2).unwrap();
+
+        // create an empty directory
+        fs_extra::dir::create_all(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder3"), false).unwrap();
+
+        // create file list
+        let mut file_list = Vec::new();
+        file_list.push("folder1/folder1_1/test1.txt".to_string());
+        file_list.push("folder2/test2.txt".to_string());
+        file_list.push("folder3/".to_string());
+
+        // run copy and expect an error
+        let result = fs::copy(file_list, &Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src"), &Path::new(TEST_FOLDER).join("test_copy_nested_folder2/target"));
+        assert!(result.is_ok());
+
+        // check target files exist
+        let path1_target = Path::new(TEST_FOLDER).join("test_copy_nested_folder2/target/folder1/folder1_1/test1.txt");
+        let path2_target = Path::new(TEST_FOLDER).join("test_copy_nested_folder2/target/folder2/test2.txt");
+
+        assert!(path1_target.exists());
+        assert!(path2_target.exists());
+
+        // check target empty directory
+        assert!(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/target/folder3").exists());
+        assert!(Path::new(TEST_FOLDER).join("test_copy_nested_folder2/target/folder3").is_dir());
+
+        // compare contents
+        let path1_src = Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder1/folder1_1/test1.txt");
+        let path2_src = Path::new(TEST_FOLDER).join("test_copy_nested_folder2/src/folder2/test2.txt");
         assert!(file_eq(path1_src, path1_target).unwrap());
         assert!(file_eq(path2_src, path2_target).unwrap());
     })
